+++ jlegmed-core/src/main/java/io/jexxa/jlegmed/core/flowgraph/ProcessingStats.java
@@ -2,4 +2,7 @@ package io.jexxa.jlegmed.core.flowgraph;
 
 import java.math.BigInteger;
 
-public record ProcessingStats(BigInteger processingErrorCount, BigInteger handledProcessingErrors, BigInteger unhandledProcessingErrors) { }
+public record ProcessingStats(BigInteger forwardedMessages,
+                              BigInteger processingErrorCount,
+                              BigInteger handledProcessingErrors,
+                              BigInteger unhandledProcessingErrors) { }