+++ jlegmed-core/src/main/java/io/jexxa/jlegmed/core/pipes/OutputPipe.java
@@ -1,10 +1,15 @@
 package io.jexxa.jlegmed.core.pipes;
 
+import java.math.BigInteger;
+
 import static io.jexxa.adapterapi.invocation.InvocationManager.getInvocationHandler;
 import static io.jexxa.common.facade.logger.SLF4jLogger.getLogger;
+import static java.math.BigInteger.valueOf;
 
 public class OutputPipe<T> {
     private InputPipe<T> inputPipe;
+    private BigInteger forwardedMessages = valueOf(0);
+
 
     public void connectTo(InputPipe<T> inputPipe)
     {
@@ -20,6 +25,10 @@ public class OutputPipe<T> {
         return inputPipe != null;
     }
 
+    public BigInteger forwardedMessages() {
+        return forwardedMessages;
+    }
+
     private void forwardToSuccessor(T data)
     {
         if (data == null) {
@@ -31,8 +40,8 @@ public class OutputPipe<T> {
             getLogger(OutputPipe.class).debug("No input pipe connected");
             return;
         }
-
         inputPipe.receive(data);
+        forwardedMessages = forwardedMessages.add(valueOf(1));
     }
 
     protected InputPipe<T> inputPipe() {